关于RooInit_NumericType_BBScriptHandler的说明
参照了ET的NumericComponent组件，注意所有注册的数值都是long类型，
Gravity、MaxFall等数值需要 / 1000才是真实数值

+++ Unity/Assets/Scripts/Codes/HotfixView/Client/Demo/Timeline/BBScript/Handler/RootInit/RootInit_NumericType_BBScriptHandler.cs
@@ -26,7 +26,7 @@ namespace ET.Client
                 return Status.Failed;
             }
             
-            int.TryParse(match.Groups["Value"].Value, out int value);
+            long.TryParse(match.Groups["Value"].Value, out long value);
             timelineComponent.RegistParam(match.Groups["NumericType"].Value, value);
             
             await ETTask.CompletedTask;
